Disable native_coverage for webrtc
Bug: http://b/116873221

Linking fails at O0 because VTable for InStream and OutStream (defined
in webrtc/common_types.h) is not found.  The fix is to mark the Rewind
function in both classes as pure virtual.

It's better to disable native_coverage here rather than pick up a
downstream change to an external codebase.

Test: Build with coverage
Change-Id: I13375f8fa02c2e9330d30f091f069518e2908f51